ChartLoader/src/requirements.txt
fastapi
uvicorn
tinkoff-investments
numpy
scipy